backend/pyproject.toml
[project]
name = "immosimple-agent"
version = "0.1.0"
description = "ImmoSimple AI agent — FastAPI + LangGraph + Claude"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "supabase>=2.7",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=1.1.14",
    "langchain-google-genai>=4.2.2",
]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = ["ruff>=0.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/immosimple_agent"]
